Move experiments into subdirectories
