+++ docs/src/app/(private)/experiments/accordion/accordion-animations.tsx
