+++ docs/src/app/(private)/experiments/accordion/accordion.horizontal.module.css
