+++ docs/src/app/(private)/experiments/collapsible/collapsible-cls.tsx
