+++ docs/src/app/(private)/experiments/collapsible/collapsible-framer.tsx
