+++ docs/src/app/(private)/experiments/collapsible/collapsible-hidden-until-found.tsx
