+++ docs/src/app/(private)/experiments/collapsible/collapsible.module.css
