+++ docs/src/app/(private)/experiments/collapsible/collapsible.tsx
