+++ docs/src/app/(private)/experiments/scroll-area/scroll-area-inset.tsx
